debian/patches/atari/: 0001-xwayland-window-manager-Make-clients
Clients should, before mapping, to a list of atoms indicating the
functional type of the window, with _NET_WM_WINDOW_TYPE_NORMAL
indicating a normal/regular top level window. Some of the clients
do not, resulting in a floating window, but with correct window dimensions.

Force them to be maximized (instead of full-screen) to easily figure
when that happens and to avoid any full-screen issues I've seen
happening on some occasions.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>